add casteljau's spline algorithm

+++ curves/src/main/java/com/nocmok/opengl/curve/curve_drawer/Grid.java
@@ -2,5 +2,5 @@ package com.nocmok.opengl.curve.curve_drawer;
 
 public interface Grid {
 
-    void setPixel(int x, int y);
+    void setPixel(double x, double y);
 }
